feat: major overhaul removing hard deps
