+++ cmd/lambda/main.go
@@ -4,11 +4,7 @@ import (
 	"os"
 
 	"github.com/aws/aws-lambda-go/lambda"
-
-	"github.com/kjansson/defcon"
-	"github.com/kjansson/yac-p/v2/pkg/controller"
-	"github.com/kjansson/yac-p/v2/pkg/prom"
-	"github.com/kjansson/yac-p/v2/pkg/types"
+	//"github.com/kjansson/yac-p/v2/pkg/controller"
 )
 
 func main() {
@@ -17,7 +13,7 @@ func main() {
 
 func HandleRequest() {
 
-	config := &types.Config{
+	config := Config{
 		ConfigFileLoader: GetS3Loader(), // Use the S3 loader for Lambda implementation
 		RemoteWriteURL:   os.Getenv("PROMETHEUS_REMOTE_WRITE_URL"),
 		AuthType:         os.Getenv("AUTH_TYPE"),
@@ -29,12 +25,7 @@ func HandleRequest() {
 		AWSRoleARN:       os.Getenv("AWS_ROLE_ARN"),
 	}
 
-	err := defcon.CheckConfigStruct(config) // Check the config struct for required fields
-	if err != nil {
-		panic(err)
-	}
-
-	c, err := controller.NewController(*config) // Create a new controller instance
+	c, err := NewController(config) // Create a new controller instance
 	if err != nil {
 		panic(err)
 	}
@@ -57,7 +48,7 @@ func HandleRequest() {
 
 	c.Logger.Log("debug", "Processing metrics")
 	// Process the metrics into timeseries format
-	timeSeries, err := prom.ProcessMetrics(metrics, c.Logger)
+	timeSeries, err := c.ConvertMetrics(metrics)
 	if err != nil {
 		panic(err)
 	}
