+++ pkg/persister/prom/prometheus_test.go
@@ -4,13 +4,11 @@ import (
 	"fmt"
 	"net/http"
 	"net/http/httptest"
+	"os"
 	"testing"
 
 	"github.com/kjansson/yac-p/v2/pkg/logger"
-	"github.com/kjansson/yac-p/v2/pkg/types"
-	io_prometheus_client "github.com/prometheus/client_model/go"
 	"github.com/prometheus/prometheus/prompb"
-	"google.golang.org/protobuf/proto"
 )
 
 func checkHeaders(r *http.Request) error {
@@ -26,25 +24,6 @@ func checkHeaders(r *http.Request) error {
 	return nil
 }
 
-func createTestMetricsFamily() []*io_prometheus_client.MetricFamily {
-	metric := []*io_prometheus_client.Metric{{
-		Label: []*io_prometheus_client.LabelPair{
-			//	{Name: proto.String("__name__"), Value: proto.String("test_gauge")},
-			{Name: proto.String("label1"), Value: proto.String("value1")},
-		},
-		Gauge: &io_prometheus_client.Gauge{
-			Value: proto.Float64(1.0),
-		},
-	}}
-
-	return []*io_prometheus_client.MetricFamily{{
-		Name:   proto.String("test_gauge"),
-		Help:   proto.String("This is a test gauge"),
-		Type:   io_prometheus_client.MetricType_GAUGE.Enum(),
-		Metric: metric,
-	}}
-}
-
 func createTestTimeSeries() []prompb.TimeSeries {
 	return []prompb.TimeSeries{
 		{
@@ -56,41 +35,6 @@ func createTestTimeSeries() []prompb.TimeSeries {
 				{Value: 1.0, Timestamp: 1234567890},
 			},
 		},
-	}
-}
-
-func TestMetricsProcessing(t *testing.T) {
-	logger := &logger.SlogLogger{}
-	err := logger.Init(types.Config{})
-	if err != nil {
-		t.Fatalf("Failed to initialize logger: %v", err)
-	}
-
-	timeseries, err := ProcessMetrics(createTestMetricsFamily(), logger)
-	if err != nil {
-		t.Fatalf("Failed to process metrics: %v", err)
-	}
-
-	for _, ts := range timeseries {
-		if len(ts.Labels) == 0 {
-			t.Fatalf("Timeseries has no labels")
-		}
-		if ts.Labels[0].Name != "__name__" {
-			t.Fatalf("Timeseries does not have __name__ label")
-		}
-		if ts.Labels[0].Value != "test_gauge" {
-
-			t.Fatalf("Timeseries does not have correct __name__ label value")
-		}
-		if len(ts.Samples) == 0 {
-			t.Fatalf("Timeseries has no samples")
-		}
-		if ts.Samples[0].Value != 1.0 {
-			t.Fatalf("Timeseries does not have correct sample value")
-		}
-		if ts.Samples[0].Timestamp == 0 {
-			t.Fatalf("Timeseries does not have correct timestamp")
-		}
 	}
 }
 
@@ -109,8 +53,12 @@ func TestMetricsPersistingNoAuth(t *testing.T) {
 		}
 	}))
 
-	logger := &logger.SlogLogger{}
-	err := logger.Init(types.Config{})
+	logger, err := logger.NewLogger(
+		os.Stdout,
+		"debug",
+		"text",
+		false,
+	)
 	if err != nil {
 		t.Fatalf("Failed to initialize logger: %v", err)
 	}
@@ -147,8 +95,12 @@ func TestMetricsPersistingBasicAuth(t *testing.T) {
 		}
 	}))
 
-	logger := &logger.SlogLogger{}
-	err := logger.Init(types.Config{})
+	logger, err := logger.NewLogger(
+		os.Stdout,
+		"debug",
+		"text",
+		false,
+	)
 	if err != nil {
 		t.Fatalf("Failed to initialize logger: %v", err)
 	}
@@ -184,8 +136,12 @@ func TestMetricsPersistingTokenAuth(t *testing.T) {
 		}
 	}))
 
-	logger := &logger.SlogLogger{}
-	err := logger.Init(types.Config{})
+	logger, err := logger.NewLogger(
+		os.Stdout,
+		"debug",
+		"text",
+		false,
+	)
 	if err != nil {
 		t.Fatalf("Failed to initialize logger: %v", err)
 	}
